refactor: Phase 2 Task 5 - SOLID refactoring of SmartMessageRenderer
🎯 PHASE 2 TASK 5 COMPLETE: SOLID Refactoring (God Object Split)

## ✅ Architecture Improvements

### Single Responsibility Principle (SRP) Applied
Split SmartMessageRenderer (468 lines, 4+ responsibilities) into 3 specialized files:

1. **message_parser_utils.dart** (77 lines)
   - Utility class for message parsing and format detection
   - Static-only interface (private constructor pattern)
   - Exports:
     * documentBlockRegex: RegExp for <document> blocks
     * decodeHtmlEntities(): HTML entity decoding
     * isRailOperationDocument(): Format detection

2. **document_proposal_card.dart** (300 lines)
   - Interactive widget for AI document proposals
   - Three actions: Validate & Save, Refine, Reject
   - Static state persistence (_validatedDocs, _dismissedDocs)
   - JSON auto-wrapping for pure JSON content
   - Uses CodeElementBuilder for syntax highlighting

3. **smart_message_renderer.dart** (216 lines, down from 468)
   - Router only: User vs AI messages, document proposals vs plain text
   - Delegates parsing to MessageParserUtils
   - Delegates UI rendering to DocumentProposalCard
   - Clean separation of concerns

## 🧪 Test Results

- ✅ Flutter analyze: No issues found (3 files)
- ✅ All 929 tests passing (2 skipped)
- ✅ Test execution time: 43 seconds
- ✅ Behavior maintained exactly as before refactoring

## 📝 Technical Details

**Files Modified:**
- src/client/lib/features/chat/presentation/widgets/smart_message_renderer.dart (468 → 216 lines)

**Files Created:**
- src/client/lib/features/chat/presentation/widgets/message_parser_utils.dart (77 lines)
- src/client/lib/features/chat/presentation/widgets/document_proposal_card.dart (300 lines)

**Benefits:**
- Single Responsibility Principle (each file has one clear purpose)
- Better maintainability (smaller, focused files)
- Easier testing (isolated logic and UI)
- Clear delegation pattern (router → utilities → components)
- Comprehensive DartDoc for all public APIs
